Started simple shared middleware
