+++ services/base-service/models/middleware.js
@@ -1,0 +1,33 @@
+// Custom middleware to check Content-Type
+function json_only(req, res, next) {
+  const contentType = req.get('Content-Type');
+  if (contentType === 'application/json' || contentType === 'application/ld+json') {
+    // Content-Type is acceptable
+    return next();
+  } else {
+    // Content-Type is not acceptable
+    res.status(400).json({ error: 'Invalid Content-Type. Only application/json or application/ld+json is allowed.' });
+  }
+}
+
+function user_can_edit(req, res, next) {
+  let expertId = `expert/${req.params.expertId}`;
+  if (!req.user) {
+    return res.status(401).send('Unauthorized');
+  }
+  if ( req.user?.roles?.includes('admin')) {
+    return next();
+  }
+
+  if( expertId === req.user.expertId ) {
+    return next();
+  }
+
+  return res.status(403).send('Not Authorized');
+}
+
+// export this middleware functions
+module.exports = {
+  json_only,
+  user_can_edit
+};
